+++ cds-feature-n8n/src/main/java/sap/capire/n8n_plugin/services/ConsoleN8NWebhookService.java
@@ -21,7 +21,7 @@ public class ConsoleN8NWebhookService extends N8nWebhookService {
   private final AtomicInteger counter = new AtomicInteger(0);
 
   public ConsoleN8NWebhookService() {
-    super("http://console", "", null);
+    super("http://console", "", Collections.emptyMap(), null);
   }
 
   /**